+++ Plugins/TableProPluginKit/CSVDialect.swift
@@ -17,6 +17,7 @@ public struct CSVDialect: Equatable, Sendable {
 
     public var delimiter: UInt8
     public var quoteChar: UInt8
+    public var escapeChar: UInt8
     public var encoding: String.Encoding
     public var lineEnding: LineEnding
     public var hasBom: Bool
@@ -30,6 +31,7 @@ public struct CSVDialect: Equatable, Sendable {
     ) {
         self.delimiter = delimiter
         self.quoteChar = quoteChar
+        self.escapeChar = quoteChar
         self.encoding = encoding
         self.lineEnding = lineEnding
         self.hasBom = hasBom